[#1662 (5)] Fixes for "product", docker building & better high time handling

+++ golang/cmd/kafka-to-postgresql-v2/postgresql/analytics-state.go
@@ -24,10 +24,10 @@ func (c *Connection) InsertStateAdd(msg *sharedStructs.StateAddMessage, topic *s
 	var cmdTag pgconn.CommandTag
 	cmdTag, err = tx.Exec(ctx, `
 		UPDATE state
-		SET    end_time = to_timestamp($2 / 1000.0)
+		SET    end_time = to_timestamp($2::BIGINT / 1000.0)
 		WHERE  asset_id = $1
 			   AND end_time IS NULL
-			   AND start_time < to_timestamp($2 / 1000.0) 
+			   AND start_time < to_timestamp($2::BIGINT / 1000.0) 
 	`, int(assetId), msg.StartTimeUnixMs)
 
 	if err != nil {
@@ -41,7 +41,6 @@ func (c *Connection) InsertStateAdd(msg *sharedStructs.StateAddMessage, topic *s
 	}
 
 	// Insert state
-	// TODO: Error if there is already a state with the same start time
 	cmdTag, err = tx.Exec(ctx, `
 		INSERT INTO state
             (
@@ -52,7 +51,7 @@ func (c *Connection) InsertStateAdd(msg *sharedStructs.StateAddMessage, topic *s
             VALUES
             (
                         $1,
-                        to_timestamp($2/1000.0),
+                        to_timestamp($2::BIGINT  / 1000.0),
                         $3
             )
 		on conflict
@@ -96,8 +95,8 @@ func (c *Connection) OverwriteStateByStartEndTime(msg *sharedStructs.StateOverwr
 	cmdTag, err = tx.Exec(ctx, `
 		DELETE FROM state
 		WHERE  asset_id = $1
-			   AND start_time >= to_timestamp($2 / 1000.0)
-			   AND start_time <= to_timestamp($3 / 1000.0) 
+			   AND start_time >= to_timestamp($2::BIGINT / 1000.0)
+			   AND start_time <= to_timestamp($3::BIGINT / 1000.0) 
 	`, int(assetId), msg.StartTimeUnixMs, msg.EndTimeUnixMs)
 
 	if err != nil {
@@ -113,10 +112,10 @@ func (c *Connection) OverwriteStateByStartEndTime(msg *sharedStructs.StateOverwr
 	// Check for overlapping state and modify it's end time
 	cmdTag, err = tx.Exec(ctx, `
 		UPDATE state
-		SET    end_time = to_timestamp($2 / 1000.0)
+		SET    end_time = to_timestamp($2::BIGINT / 1000.0)
 		WHERE  asset_id = $1
-			   AND end_time > to_timestamp($2 / 1000.0)
-			   AND end_time <= to_timestamp($3 / 1000.0) 
+			   AND end_time > to_timestamp($2::BIGINT / 1000.0)
+			   AND end_time <= to_timestamp($3::BIGINT / 1000.0) 
 	`, int(assetId), msg.StartTimeUnixMs, msg.EndTimeUnixMs)
 
 	if err != nil {
@@ -132,10 +131,10 @@ func (c *Connection) OverwriteStateByStartEndTime(msg *sharedStructs.StateOverwr
 	// Check for overlapping state and modify it's start time
 	cmdTag, err = tx.Exec(ctx, `
 		UPDATE state
-		SET    start_time = to_timestamp($3 / 1000.0)
+		SET    start_time = to_timestamp($3::BIGINT / 1000.0)
 		WHERE  asset_id = $1
-			   AND start_time >= to_timestamp($2 / 1000.0)
-			   AND start_time < to_timestamp($3 / 1000.0) 
+			   AND start_time >= to_timestamp($2::BIGINT / 1000.0)
+			   AND start_time < to_timestamp($3::BIGINT / 1000.0) 
 	`, int(assetId), msg.StartTimeUnixMs, msg.EndTimeUnixMs)
 
 	if err != nil {
@@ -157,8 +156,8 @@ func (c *Connection) OverwriteStateByStartEndTime(msg *sharedStructs.StateOverwr
              end_time,
              state)
 		VALUES      ($1,
-					 to_timestamp($2 / 1000.0),
-					 to_timestamp($3 / 1000.0),
+					 to_timestamp($2::BIGINT / 1000.0),
+					 to_timestamp($3::BIGINT / 1000.0),
 					 $4) 
 	`, int(assetId), msg.StartTimeUnixMs, msg.EndTimeUnixMs, int(msg.State))
